Store points instead of "current_value"
points() was showing up frequently in a flamegraph, likely due to the
fact that it required doing math to calculate.

This commit doubles the speed of the simulation by storing the point
value on the Die instead of the current_value. Since the simulator only
actually cares about the point value and not the current_value of the
Die, this prevents having to do a ton of math in hot loops.